+++ externals/coda-oss/.github/workflows/build_unittest.yml
@@ -36,8 +36,7 @@ jobs:
         cd ..
         cd target-Debug
         cmake --build . --config Debug -j
-    - name: test
-    # should run w/o install
+    - name: test # should run w/o install
       run: |
         cd target-Release
         ctest -C Release --output-on-failure
@@ -52,17 +51,43 @@ jobs:
         cd target-Debug
         cmake --build . --config Debug --target install
         cd ..
+
+  build-msbuild-windows:
+    strategy:
+      matrix:
+        os: [windows-latest]
+        platform: [x64]
+        configuration: [Debug] # Debug turns on more compiler warnings
+    name: ${{ matrix.os }}-msbuild
+    runs-on: ${{ matrix.os }}
+
+    steps:
+    - uses: actions/checkout@v3
+    - name: configure
+      run: |
+        ls env:
+        mkdir out
+        cd out
+        cmake .. -DCMAKE_INSTALL_PREFIX=install\${{ matrix.platform }}-${{ matrix.configuration }} -DENABLE_PYTHON=OFF
+    - name: build
+      run: |
+        cd out
+        cmake --build . --config ${{ matrix.configuration }} -j
+        cmake --build . --config ${{ matrix.configuration }} --target install
     - name: Add msbuild to PATH
-      uses: microsoft/setup-msbuild@v1.1
+      uses: microsoft/setup-msbuild@v1 # https://github.com/marketplace/actions/setup-msbuild
       with:
         msbuild-architecture: x64
     - name: msbuild
       run: |
-        msbuild coda-oss.sln /p:configuration=Release
-        msbuild coda-oss.sln /p:configuration=Debug
-    #- name: mstest
-    #  run: |
-        # mstest /testcontainer:${RUNNER_WORKSPACE}\coda-oss\x64\Debug\UnitTest.dll
+        msbuild coda-oss.sln /p:configuration=${{ matrix.configuration }}
+    # Can't figure out how to make this work :-(
+    #- name: vstest 
+    #  uses: microsoft/vstest-action@v1.0.0 # https://github.com/marketplace/actions/vstest-action
+    #  with:
+    #      testAssembly: UnitTest.dll
+    #      searchFolder: D:\a\nitro\nitro\${{ matrix.platform }}\${{ matrix.configuration }}
+    #      runInParallel: true  
 
   build-linux-cmake:
     strategy:
