+++ externals/coda-oss/modules/c++/coda-oss.vcxproj.filters
@@ -963,6 +963,12 @@
     <ClInclude Include="hdf5.lite\include\hdf5\lite\H5_.h">
       <Filter>hdf5.lite</Filter>
     </ClInclude>
+    <ClInclude Include="mt\include\import\mt.h">
+      <Filter>mt</Filter>
+    </ClInclude>
+    <ClInclude Include="include\UnitTest.h">
+      <Filter>include</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="pch.cpp" />
@@ -1586,5 +1592,8 @@
       <Filter>std</Filter>
     </None>
     <None Include="coda-oss.ruleset" />
+    <None Include="sys\source\CppUnitTestAssert_.cpp_">
+      <Filter>sys</Filter>
+    </None>
   </ItemGroup>
 </Project>
