+++ externals/coda-oss/modules/c++/sys/source/FileFinder.cpp
@@ -208,6 +208,17 @@ std::vector<std::string> sys::FileFinder::search(
     return files;
 }
 
+static fs::path parent_path(const fs::path& p)
+{
+    // If the parent_path() is the same, we've reached to root.
+    const auto retval = p.parent_path();
+    if (retval.empty() || (retval == p))
+    {
+        throw std::runtime_error("At root of filesystem: " + p.string());
+    }
+    return retval;
+}
+
 static fs::path findFirst(const sys::FilePredicate& pred, const fs::path& startingDirectory)
 {
     auto dir = startingDirectory;
@@ -228,7 +239,8 @@ static fs::path findFirst(const sys::FilePredicate& pred, const fs::path& starti
         {
             throw std::logic_error("Won't traverse above .git directory at: " + dir.string());
         }
-        dir = dir.parent_path();
+
+        dir = parent_path(dir);
     }
 }
 fs::path sys::findFirstFile(const fs::path& startingDirectory, const fs::path& filename)
@@ -297,7 +309,7 @@ fs::path sys::test::findRootDirectory(const fs::path& p, const std::string& root
     // Once we're at a .git directory, we have to go down, not up ... or fail.
     if (!is_directory(p / ".git"))
     {
-        return findRootDirectory(p.parent_path(), rootName, isRoot);
+        return findRootDirectory(parent_path(p), rootName, isRoot);
     }
     
     // TODO: since we're in the "FileFinder" module, maybe try a bit harder to find "rootName"?
@@ -316,7 +328,7 @@ static inline std::string Platform()
 
 static fs::path findCMakeRoot(const fs::path& path, const fs::path& dir)
 {
-   	static const auto platform_and_configuration = ::Platform() + "-" + ::Configuration(); // "x64-Debug"
+    static const auto platform_and_configuration = ::Platform() + "-" + ::Configuration(); // "x64-Debug"
     const auto pred = [&](const fs::path& p)
     {
         if (p.filename() == platform_and_configuration)
@@ -338,7 +350,7 @@ static fs::path findCMakeRoot(const fs::path& path, const fs::path& dir)
 
         return false;
     };
-	return sys::test::findRootDirectory(path, "", pred);
+    return sys::test::findRootDirectory(path, "", pred);
 }
 
 fs::path findCMake_Root(const fs::path& path,
@@ -404,13 +416,8 @@ static fs::path find_dotGITDirectory_(const fs::path& p, const fs::path& initial
     {
         return p;
     }
-
-    auto parent = p.parent_path();
-    if (parent.empty())
-    {
-        throw std::invalid_argument("Can't find .git/ anywhere in: " + initial.string());    
-    }
-    return find_dotGITDirectory_(parent, initial);
+    
+    return find_dotGITDirectory_(parent_path(p), initial);
 }
 fs::path sys::test::find_dotGITDirectory(const fs::path& p)
 {
