+++ externals/coda-oss/modules/c++/xml.lite/source/Serializable.cpp
@@ -26,8 +26,10 @@ void xml::lite::Serializable::serialize(io::OutputStream& os)
 {
     xml::lite::Element *root = getDocument()->getRootElement();
     if (root != nullptr)
+    {
         os.write("<?xml version=\"1.0\" encoding=\"utf-8\"?>\n");
-    root->print(os);
+        root->print(os);
+    }
 }
 
 void xml::lite::Serializable::deserialize(io::InputStream& is)
